Added the registration logs.
Tweaked the h1 CSS for the Control Panel in Cosora.
Improved the support for autocompleters.
Reverted some of the int64s back to ints.
The per-request user variable should now have the IPs for guests.
Removed some obsolete statements.
Removed a couple obsolete phrases.
Moved a couple more accumulator initializers out of the store initializers.
Refactored one of the report statements to use a store instead of a statement.
Created the menuhead CSS class.
Added the GetOffset method to the LogStore interface and refactored the mod and admin log pages to use it.
Changed the labels on the adminlog and modlog pages to make them more understandable.
Removed the generated templates from Git.

Run the patcher / update script to get the new table.

+++ common/site.go
@@ -52,7 +52,7 @@ type config struct {
 	SslPrivkey   string
 	SslFullchain string
 
-	MaxRequestSize     int64
+	MaxRequestSize     int
 	CacheTopicUser     int
 	UserCacheCapacity  int
 	TopicCacheCapacity int
